Move logic to presenter
